refactor: 팀원 모집 공통 코드 및 스타일 정리
