+++ src/components/Team/components/NotificationCard/NotificationCard.module.scss
@@ -46,7 +46,6 @@
     flex-shrink: 0;
     color: #a8a8a8;
     font-size: 10px;
-    font-weight: 400;
     line-height: 160%;
     white-space: nowrap;
   }
