+++ src/components/Team/components/TeamListHeader/TeamListHeader.module.scss
@@ -31,7 +31,6 @@
     margin: 0;
     transform: translate(-50%, -50%);
     color: #000;
-    font-size: 16px;
     font-weight: 600;
     line-height: 160%;
     white-space: nowrap;
