+++ src/components/Team/components/TeamNotificationHeader/TeamNotificationHeader.module.scss
@@ -30,7 +30,6 @@
     margin: 0;
     transform: translate(-50%, -50%);
     color: #000;
-    font-size: 16px;
     font-weight: 600;
     line-height: 160%;
     white-space: nowrap;
@@ -64,7 +63,6 @@
     padding: 8px;
     color: #000;
     font-size: 12px;
-    font-weight: 400;
     line-height: 160%;
     text-align: center;
     white-space: nowrap;
