+++ src/components/Team/components/TeamSearchBar/TeamSearchBar.module.scss
@@ -23,7 +23,6 @@
   color: #727272;
   font-family: inherit;
   font-size: 12px;
-  font-weight: 400;
   line-height: 160%;
   outline: none;
