+++ src/pages/team/notifications/TeamNotificationsPage.module.scss
@@ -8,6 +8,7 @@
 .list {
   display: flex;
   flex-direction: column;
+  gap: 12px;
   padding-top: 12px;
 }
 
@@ -16,7 +17,6 @@
   padding: 12px 20px;
   color: #727272;
   font-size: 12px;
-  font-weight: 400;
   line-height: 160%;
 }
 
@@ -57,7 +57,6 @@
   padding: 0 20px;
   color: #a8a8a8;
   font-size: 14px;
-  font-weight: 400;
   line-height: 100%;
   text-align: center;
 }
